fix: test e2e peer store relies on hashset order
testswitch can fail with the following message:

```
Check failed: storedInfo2.protos.toSeq() == switch1.peerInfo.protocols
    storedInfo2.protos.toSeq() was @["/test/proto/1.0.0", "/ipfs/id/1.0.0"]
    switch1.peerInfo.protocols was @["/ipfs/id/1.0.0", "/test/proto/1.0.0"]
```

The test assumes a specific order of items in a hashset.
This commit converts the sequence into a hashset fixing this issue locally.

Is there a reason for PeerInfo managing supported protocols in a seq, while PeerStore manages it in a HashSet?