fix: remove returned Futures from switch.start
The proc `start` returned a seq of futures that was mean to be awaited by the caller. However, the start proc itself awaited each Future before returning it, so the ceremony requiring the caller to await the Future, and returning the Futures themselves, was unnecessary.

Remove `switch.start` return type (implicit `Future[void]`)

Update tutorials and examples to reflect the change.