feat: Add callback for pessimistic transport start
Add switch.listenError callback that is called when a transport fails to start. This is a pessimistic approach meaning that it will fail on the first address that fails to start in the transport, which has been discussed in #662. If we wanted to be optimistic inside of the transport.start (a la #661), we can add that in a future commit.