move `allFutureThrowing` helper to tests
We use a generically seeming helper to help us track three `Future`
exceptions throughout the entire codebase. Move the helper to tests
where it finds more use. The helper has several design flaws that make
it incompatible with `{.async: (raises).}` usage, and has the potential
to swallow one of the three exceptions that we track with it.

The semantics should not change in this PR, this is a pure refactoring.