annotate `upgrademngrs` with `{.async: (raises).}`
The next layer of modules that can get `{.async: (raises).}` annotations
are the `upgrademngrs`. This is the first layer that mixes so many error
types that it makes sense to start with `{.async: (raises: [LPError]).}`
especially because the callers add raw `LPError` as well. It may be
worthwhile to try and wrap inner exceptions in a proper outer type,
but as this pattern spans large parts of the library, such wrapping
should be investigated separately from initial annotation work.